Extract AvatarMenu into its own component
